refactor : component index명  수정
