+++ client/src/pages/qna/index.tsx
@@ -1,4 +1,4 @@
-import Layout from '@/components/Layout/Layout';
+import Layout from '@/components/Layout';
 import React from 'react';
 
 export default function Qna() {
